+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bug Fixes
 1. [#81](https://github.com/influxdata/influxdb-client-csharp/pull/81): Fixed potentially hangs on of WriteApi.Dispose()
+1. [#82](https://github.com/influxdata/influxdb-client-csharp/pull/82): Avoid to deadlock when is client used in UI
 
 ## 1.7.0 [2020-04-17]